requirements.txt
torch
transformers==4.43.3
accelerate
datasets
evaluate
matplotlib
pandas
peft
scikit-learn
seaborn
trl
wandb
dill
tqdm
python-dotenv
openai
streamlit
together